requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
etils==1.5.2
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
flask==3.0.3
flatbuffers==24.3.25
flax==0.8.4
fonttools==4.53.0
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
install==1.3.5
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jupyter-client==8.6.2
jupyter-core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.1
numpy==1.26.4
nvidia-pyindex==1.0.9
opencv-python==4.9.0.80
openvino==2024.1.0
openvino-dev==2024.1.0
openvino-telemetry==2024.1.0
opt-einsum==3.3.0
optax==0.2.2
optree==0.11.0
orbax-checkpoint==0.5.15
packaging==23.2
pandas==2.2.2
parso==0.8.4
pika==1.3.2
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.45
protobuf==3.20.1
psutil==5.9.8
pure-eval==0.2.2
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-decision-forests==1.8.1
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflowjs==4.19.0
tensorstore==0.1.60
termcolor==2.4.0
tf-keras==2.16.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing-extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Wand==0.6.13
wcwidth==0.2.13
werkzeug==3.0.3
wrapt==1.16.0
wurlitzer==3.1.0
yml==0.0.1
zipp==3.19.1
